backend/requirements.txt
# FastAPI & ASGI
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
aiomysql==0.2.0
alembic==1.13.1
pymysql==1.1.0

# Pydantic & Settings
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Redis & Caching
redis==5.0.4
aioredis==2.0.1

# Email
fastapi-mail==1.4.1
jinja2==3.1.4

# Rate Limiting
slowapi==0.1.9

# File Storage
aiofiles==23.2.1
python-magic==0.4.27

# Background Tasks
celery==5.3.6
flower==2.0.1

# HTTP Client
httpx==0.27.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-httpx==0.30.0
anyio==4.4.0

# Dev utilities
python-dotenv==1.0.1
structlog==24.1.0
